requirements.txt
build==1.2.2.post1
click==8.1.8
JCC
joblib==1.4.2
lucene
nltk==3.9.1
packaging==24.1
pyproject_hooks==1.2.0
regex==2024.11.6
setuptools==75.2.0
textblob==0.19.0
tqdm==4.67.1
whoosh
PyYAML
psycopg2
requests
pandas
matplotlib.pyplot
seaborn